contibutions were put to separate component
(cherry picked from commit abd4a37829bae3b427a3e8364c54894318c55582)
